rename and move fake connector
